extract serializer
